restructure cache test cases
